[Refactor] 캘린더 날짜 선택 리펙토링
