+++ src/features/calendar/index.ts
@@ -1,5 +1,5 @@
 export { Calendar } from './ui/calendar/Calendar'
-export { PickDate } from './ui/PickDate'
+export { PickDate } from './ui/pickDate/PickDate'
 export { DateListOverlay } from './ui/overlay/DateListOverlay'
 
 export { useSelectedDate } from './model/useSelectedDate'
